Make ReactMqttClient more robust against unexpected errors
The currently existing E_DEPRECATED in the reactphp Promise code surfaced an
interesting bug: When you run the ReactMqttClient in an environment that e.g.
throws that E_DEPRECATED as an exception `->otherwise` chains may not be
executed, because the promise-callback-chains are not terminated with `done()`.

This can (currently) be simulated by setting an aggressive error-handler in the
testsuite:

```
set_error_handler(function ($_type, $message) {
   throw new Error($message);
});
```

Testcase that rely on promise rejections (e.g. test_connect_failure) will hang
until the test times out.

This commit fixes all non-returned callback-chains with a `done()`-termination
as it is best-practice.